Doing one of my courses I learnt about segmented control and it occurred to me that it could work very well on my cheque writing app.
I took away the existing buttons and implemented the segmented control for the different languages which was super easy. Only thing now is that it is not 100% intuitive as it does not translate anything until you change language. Either I leave it as it is or I enable live translation as you type which is a completely different beast. I'm leaning towards the latter, but that's for another day.